Hosting metrics: abstract useGroupByTime to create useSiteMetrics400vs500Data line chart
