+++ client/my-sites/site-monitoring/components/site-monitoring-bar-chart/use-metrics-bar-chart-data.ts
@@ -1,7 +1,7 @@
 import { getLocaleSlug } from 'i18n-calypso';
-import { useMemo } from 'react';
+import { useGroupByTime } from '../../hooks/use-group-by-time';
 import { TimeRange } from '../../metrics-tab';
-import { PeriodData, useSiteMetricsQuery } from '../../use-metrics-query';
+import { useSiteMetricsQuery } from '../../use-metrics-query';
 
 const STATUS_CODES = [ 200, 401, 400, 404, 500 ];
 
@@ -18,35 +18,6 @@ function useSecondsWindow( timeRange: TimeRange ) {
 		return 4 * 3600;
 	}
 	return 24 * 3600;
-}
-function useGroupByTime( periods: PeriodData[], secondsWindow: number ) {
-	return useMemo( () => {
-		const groupedData: { [ key: number ]: { [ statusCode: string ]: number } } = {};
-
-		periods.forEach( ( period ) => {
-			const groupTimestamp = period.timestamp - ( period.timestamp % secondsWindow );
-			if ( ! groupedData[ groupTimestamp ] ) {
-				groupedData[ groupTimestamp ] = {};
-			}
-			// Count status codes
-			const statusCodes = Object.keys( period.dimension );
-			statusCodes.forEach( ( statusCode ) => {
-				if ( ! groupedData[ groupTimestamp ][ statusCode ] ) {
-					groupedData[ groupTimestamp ][ statusCode ] = 0;
-				}
-				groupedData[ groupTimestamp ][ statusCode ]++;
-			} );
-		} );
-
-		const dataForBarChart = [];
-		for ( const statusCode of STATUS_CODES ) {
-			dataForBarChart.push(
-				Object.values( groupedData ).map( ( group ) => group[ statusCode ] || 0 )
-			);
-		}
-		const labels = Object.keys( groupedData ) as unknown as number[];
-		return { groupedData, dataForBarChart, labels };
-	}, [ periods, secondsWindow ] );
 }
 
 function getLongDate( date: Date, locale: string ): string {
@@ -68,20 +39,22 @@ function getShortDate( date: Date, locale: string ): string {
 
 export function useMetricsBarChartData( { siteId, timeRange }: UseMetricsBarChartDataParams ) {
 	const { start, end } = timeRange;
-	const metric = 'requests_persec';
-	const dimension = 'http_status';
 	const { data } = useSiteMetricsQuery( siteId, {
 		start,
 		end,
-		metric,
-		dimension,
+		metric: 'requests_persec',
+		dimension: 'http_status',
 	} );
 
 	const secondsWindow = useSecondsWindow( timeRange );
-	const { dataForBarChart, labels } = useGroupByTime( data?.data?.periods || [], secondsWindow );
+	const { dataGroupedByTime, labels } = useGroupByTime(
+		data?.data?.periods || [],
+		secondsWindow,
+		STATUS_CODES
+	);
 
 	return {
-		data: [ STATUS_CODES.map( ( code ) => `HTTP ${ code }` ), ...dataForBarChart ] as [
+		data: [ STATUS_CODES.map( ( code ) => `HTTP ${ code }` ), ...dataGroupedByTime ] as [
 			string[],
 			...number[][]
 		],
